refactor: Rename for more readability
