+++ src/entities/Map/routes/getMapPlacesMostActive.ts
@@ -7,30 +7,18 @@ import {
   numeric,
   oneOf,
 } from "decentraland-gatsby/dist/entities/Schema/utils"
+import { flat, sort } from "radash"
 
 import { getHotScenes } from "../../../modules/hotScenes"
 import { getSceneStats } from "../../../modules/sceneStats"
 import PlaceModel from "../../Place/model"
-import { getPlaceListQuerySchema } from "../../Place/schemas"
-import {
-  FindWithAggregatesOptions,
-  GetPlaceListQuery,
-  PlaceListOrderBy,
-} from "../../Place/types"
+import { FindWithAggregatesOptions, PlaceListOrderBy } from "../../Place/types"
 import { DEFAULT_MAX_LIMIT } from "../types"
 import { placesWithCoordinatesAggregates } from "../utils"
-import { getPlaceMostActiveList } from "./getPlaceMostActiveList"
+import { validateGetPlaceListQuery } from "./getMapPlaces"
 
-export const validateGetPlaceListQuery = Router.validator<GetPlaceListQuery>(
-  getPlaceListQuerySchema
-)
-
-export const getPlaceList = Router.memo(
+export const getMapPlacesMostActive = Router.memo(
   async (ctx: Context<{}, "url" | "request">) => {
-    if (ctx.url.searchParams.get("order_by") === PlaceListOrderBy.MOST_ACTIVE) {
-      return getPlaceMostActiveList(ctx)
-    }
-
     const query = await validateGetPlaceListQuery({
       positions: ctx.url.searchParams.getAll("positions"),
       offset: ctx.url.searchParams.get("offset"),
@@ -38,12 +26,7 @@ export const getPlaceList = Router.memo(
       only_favorites: ctx.url.searchParams.get("only_favorites"),
       only_featured: ctx.url.searchParams.get("only_featured"),
       only_highlighted: ctx.url.searchParams.get("only_highlighted"),
-      order_by:
-        oneOf(ctx.url.searchParams.get("order_by"), [
-          PlaceListOrderBy.LIKE_SCORE_BEST,
-          PlaceListOrderBy.UPDATED_AT,
-          PlaceListOrderBy.CREATED_AT,
-        ]) || PlaceListOrderBy.LIKE_SCORE_BEST,
+      order_by: PlaceListOrderBy.MOST_ACTIVE,
       order:
         oneOf(ctx.url.searchParams.get("order"), ["asc", "desc"]) || "desc",
       with_realms_detail: ctx.url.searchParams.get("with_realms_detail"),
@@ -51,11 +34,26 @@ export const getPlaceList = Router.memo(
       categories: ctx.url.searchParams.getAll("categories"),
     })
 
+    const sceneStats = await getSceneStats()
+
+    const hotScenes = getHotScenes()
+
+    const hotScenesParcels = hotScenes.map((scene) => scene.parcels)
+
+    const hotScenesPositions = flat(hotScenesParcels).map((scene) =>
+      scene.join(",")
+    )
+
     const userAuth = await withAuthOptional(ctx)
 
-    if (bool(query.only_favorites) && !userAuth?.address) {
+    if (
+      (bool(query.only_favorites) && !userAuth?.address) ||
+      (numeric(query.offset) ?? 0) > hotScenes.length
+    ) {
       return new ApiResponse([], { total: 0 })
     }
+
+    const positions = new Set(query.positions)
 
     const options: FindWithAggregatesOptions = {
       user: userAuth?.address,
@@ -65,30 +63,44 @@ export const getPlaceList = Router.memo(
         DEFAULT_MAX_LIMIT,
       only_favorites: !!bool(query.only_favorites),
       only_highlighted: !!bool(query.only_highlighted),
-      positions: query.positions,
-      order_by: query.order_by,
+      positions: query.positions.length
+        ? hotScenesPositions.filter((position) => positions.has(position))
+        : hotScenesPositions,
+      order_by: PlaceListOrderBy.MOST_ACTIVE,
       order: query.order,
       search: query.search,
       categories: query.categories,
     }
 
-    const [data, total, sceneStats] = await Promise.all([
-      PlaceModel.findWithCoordinatesAggregates(options),
-      PlaceModel.countPlacesWithCoordinatesAggregates(options),
-      getSceneStats(),
-    ])
+    const { offset, limit, order, ...extraOptions } = options
+    const places = await PlaceModel.findWithCoordinatesAggregates({
+      offset,
+      limit,
+      order,
+      ...extraOptions,
+    })
 
-    const hotScenes = getHotScenes()
-
-    const response = placesWithCoordinatesAggregates(
-      data,
-      hotScenes,
-      sceneStats,
-      {
-        withRealmsDetail: !!bool(query.with_realms_detail),
-      }
+    const hotScenePlaces = sort(
+      Object.entries(
+        placesWithCoordinatesAggregates(places, hotScenes, sceneStats, {
+          withRealmsDetail: !!query.with_realms_detail,
+        })
+      ),
+      ([, place]) => place.user_count || 0,
+      !order || order === "desc"
     )
 
-    return new ApiResponse(response, { total })
+    const total = hotScenePlaces.length
+
+    const from = numeric(offset || 0, { min: 0 }) ?? 0
+    const to =
+      numeric(from + (limit || DEFAULT_MAX_LIMIT), {
+        min: 0,
+        max: DEFAULT_MAX_LIMIT,
+      }) ?? DEFAULT_MAX_LIMIT
+
+    return new ApiResponse(Object.fromEntries(hotScenePlaces.slice(from, to)), {
+      total,
+    })
   }
 )
