+++ src/entities/Map/routes/index.ts
@@ -2,7 +2,7 @@ import withCors from "decentraland-gatsby/dist/entities/Route/middleware/withCor
 import routes from "decentraland-gatsby/dist/entities/Route/wkc/routes"
 import env from "decentraland-gatsby/dist/utils/env"
 
-import { getPlaceList } from "./getPlaceList"
+import { getMapPlaces } from "./getMapPlaces"
 
 export const DECENTRALAND_URL = env("DECENTRALAND_URL", "")
 
@@ -23,5 +23,5 @@ export default routes((router) => {
       ],
     })
   )
-  router.get("/map", getPlaceList)
+  router.get("/map", getMapPlaces)
 }, {})
